fix(model)!: add a unknown variant to `PremiumType` (#1996)
This makes the size 2 bytes instead of 1, but it makes it less likely to
cause issues with serialization.

For release notes
-----------------

Add unknown variant to [`PremiumType`], this is a slight breaking change
as the type is no longer `repr(u8)` which means conversions with the
`as` keyword is no longer possible. Instead `Into` and `From` should be
used instead. ([#1996])

[`PremiumType`]: https://docs.rs/twilight-model/latest/twilight_model/user/enum.PremiumType.html
[#1996]: https://github.com/twilight-rs/twilight/pull/1996