docs(gateway)!: touch up `stream` and rename `start_` fns -> `create_` (#1981)
Renames the `start_` functions in the `stream` module to `create_` to
better reflect the fact that shards are only "started" when
`next_message` is called on them.

Additionally, `start_cluster` is renamed to `create_bucket` to better
align with the previous `ShardScheme::Bucket` naming and now has an
associated example. Also documents the fact that a shared session queue
is *not* set by default; users must specify it in the `Fn(ShardId) ->
Config` cluster.