fix(lavalink): remove server update on state update (#2025)
Remove the cache of server updates for a channel when leaving a voice channel.
When joining a channel, two halves are required to be received: a session and a
server update. A server update may not be received yet when disconnecting, so we
need to remove the cache of awaiting server updates so that rejoining a voice
channel functions.

Also fixes the intents for the Lavalink example.