feat(http, model): forum channel layouts (#2016)
Add support for forum channel layouts. This is a new enum with three
representations: "Not Set" (0); "List View" (1); and "Gallery View" (2).
The field is present on channels of the forum type as the
`default_forum_layout` field, which is optional. It can be updated via
`UpdateChannel::default_forum_layout` in the HTTP crate.

Closes #2010.