feat(model): message type deletable check methods
Add methods on `MessageType` to determine whether the message can be
deleted, as some messages can't be deleted even by server
administrators. This uses the table present on the documentation for
message types. Because this information is documented, we should expose
it. Some message types can only be deleted by people with certain
documented permissions, which at the moment is only
`AUTO_MODERATION_ACTION` which requires the `MANAGE_MESSAGES`
permission.

Included are two methods: `deletable` which doesn't check against a set
of permissions, and `deletable_by_permissions` which *does* check
against a set of permissions. Tests are included (and simplified along
with the rest of the tests).