refactor(model): abstract null boolean visitor
Abstract the `RoleTags::premium_subscriber` null boolean visitor to the
visitor module. Work is being done on adding a new field that functions
exactly like `RoleTags::premium_subscriber`, where a missing field means
false and a null value means true. It's best to abstract the visitor
built for the field so it can be re-used elsewhere.