feat(gateway)!: mark `Message` exhaustive (#2022)
The `Message` enum is the subset of the WebSocket message types relevant
to end users of the Discord gateway (`binary` frames are converted to
`text` frames by shards). It is unlikely for Discord to expand their use
of control frames and even more unlikely for it to use any currently
reserved control frames. It is therefore unnecessary to mark `Message`
as `#[non_exhaustive]`, especially since it can complicate user code.