test(model): add real-world role tag tests (#2033)
The tests used for Role Tags weren't sets of role tags one would find in
the real world. In particular, the test for "all" role tag fields can't
happen because bots can't be premium subscribers. To fix this there are
now three tests: one for bots, one for premium subscribers, and one for
no role tags.