feat(gateway): notify on disconnect (#2003)
Instead of shards transparently reconnecting they should send out a
close message and only reconnect after the gateway drops the connection,
which it does after having first sent out a close message reply. Users
are then notified of the shard disconnecting/reconnecting and can handle
it as they whish (this was previously achieved through
`ShardEvent::Disconnected`).

To achieve I had to add a new `Io` `ReceiveMessageErrorType` variant
(Tungstenite disconnects on errors) which sadly reintroduces #1428,
hence the awkward workaround for it. I plan to land a proper fix in
tokio-rustls after which the workaround can be dropped.