feat(http, model): get current authorization route
Add support for the "Get Current Authorization Information" route, at
the endpoint `GET /oauth2/@me`. This endpoint takes a client with a
bearer token and fetches the application, when the access token expires,
the authorized scopes, and the user that authorized the access token if
applicable.

Manually tested using the written example:

```
stdout:
Application: kona
User: team959818497288982639
```

Closes #1643.