feat!(model): typed locales
Adds a typed Locale struct with associated constants for all known
values, replacing the Strings used in the codebase. The Locale type has
two methods of note: `english_name` and `native_name`.

These values are taken from the Discord documentation:
<https://discord.com/developers/docs/reference#locales>

This PR is meant to show how the new raw value representation introduced
in PR #2045 looks when "typing" a value. This was previously left to the
user to know what localizations might exist and to ensure a typo isn't
made. This `Locale` type is still functionally a string in all but name
(literally, as it implements the same traits and derefs in the same way
String does), but it contains methods and the known values registered by
Twilight as "sugar" on top of it.