fix(model): zeroable forum tag emoji ids
ForumTag emoji ID fields can be an ID of "0". This is a bug on Discord's
end, but because our IDs require non-zero integers as an optimization
technique, they fail to deserialize in that case. Because the `emoji_id`
field is already wrapped in an `Option`, we can use a custom visitor to
deserialize "0" value IDs as `None`s.

Closes #1979.