feat(model)!: don't truncate bitflags
Bitflags received from the Discord API are truncated when deserialized.
However, we have recently taken a model of accepting "unknown
information", with our recent reworks of enums the shining example.
Instead of failing to deserialize / ignore "unknown variants" we have
taken the step to treat unknown data as "first class", in that Twilight
not necessarily registering known values is okay and users can manually
handle variants as they need. We should apply the same logic to bitflags
and not trim unknown data.

This unfortunately makes use of unsafe code for constructing bitflags in
tests, because `bitflags` has an unorthodox definition of what "unsafe"
is for its `Bitflags::from_bits_unchecked` function. `bitflags` treats
unknown bits as being "unsafe", and so the function to construct
bitflags with possibly unknown variants is unsafe. I have added notes
detailing this.

This is a breaking change because unknown bits are no longer truncated.

Blocks on PR #2088.