fix(gateway, http, lavalink): don't debug tokens
Tokens in the gateway crate's `Shard`, the HTTP crate's `Client` and
`ClientBuilder`, and the lavalink crate's `NodeConfig` are formatted via
default Debug implementations. This can be a security hazard for users,
as this may unknowingly print authorization tokens into their logs.
Instead of deriving default Debug implementations, take the step to
manually implement Debug formatters to redact the tokens. Tests are
included.