fix(gateway)!: only poll `MessageSender` when identified (#2002)
Fixes an issue where queued up commands are sent prior to the shard
identifying (either on startup or during a reconnect). This was possible
through splitting close frames and commands into two different channels,
which as a side benefit also causes close frames to be prioritized and
not require a ratelimit permit. Restricting `MessageSender::send` to
commands was also mirrored to `Shard::send` which should reduce
confusion of it vs `Shard::close`.

Whilst it's possible to send `Heartbeat`, `Identify` and `Resume`, this
PR also removes the `Command` trait from these commands since they're
managed by the shard and users sending them causes conflicts. The
remaining items implementing `Command` may only be sent whilst the
`Shard` is identified, as this PR now assumes.