refactor(gateway, model)!: move ShardId from gateway to model (#2097)
Shard IDs are used outside of the gateway crate, particularly in the
model crate where they are used for Ready and Identify payloads. The
gateway's `ShardId` type offers some quality of life additions over a
plain `[u64; 2]` array, such as enabling a niche optimization by using
a `NonZeroU64` for the total field.

We can enable model's usage of the type by moving `ShardId` to the model
crate and re-exporting it from the gateway.