fix(gateway-queue): skip sleep on failed messages (#2113)
In `twilight-gateway-queue` the loop for handling requests is delayed
for a bit over 5 seconds - the identify ratelimit - after sending a
message back over the oneshot `Sender` to the requester. The loop is
delayed regardless of whether the receiving half hung up. In a correct
implementation, this means that the requester holding the receiver half
that was waiting to identify will never identify.

This is a problem because when using the `gateway-queue` crate the
server will still keep the request queued and attempt to send a response
if the client cancels the request. If a cluster is started and is
running 100 shards then the queue will quickly have 100 requests
waiting. If the cluster is shutdown and restarted then all of the
previous 100 requests will need to be worked through with waiting
between them, resulting in not being able to start shards for the next
600 seconds. We can resolve this by immediately sending a message to the
next shard in the queue when an error is received.

Tokio documents the behavior of `oneshot::Sender::send`:

> A successful send occurs when it is determined that the other end of
> the channel has not hung up already. An unsuccessful send would be one
> where the corresponding receiver has already been deallocated. Note
> that a return value of Err means that the data will never be received,
> but a return value of Ok does not mean that the data will be received.
> It is possible for the corresponding receiver to hang up immediately
> after this function returns Ok.

- <https://docs.rs/tokio/latest/tokio/sync/oneshot/struct.Sender.html#method.send>

This means that it is always okay to skip sleeping and immediately send
a message to the next shard.