feat(model, util, validate)!: flatten `CommandOptionChoice` (#2081)
Refactor
`twilight_model::application::command::option::CommandOptionChoice` to
be struct rather than an enum. As an enum, `CommandOptionChoice`
contained three variants:

```rust
pub enum CommandOptionChoice {
    String(CommandOptionChoiceData<String>),
    Integer(CommandOptionChoiceData<i64>),
    Number(CommandOptionChoiceData<f64>),
}
```

`CommandOptionChoiceData`, in turn, was defined as:

```rust
pub struct CommandOptionChoiceData<T> {
    pub name: String,
    pub name_localizations: Option<HashMap<String, String>>,
    pub value: T,
}
```

Notably, `name` and `name_localizations` are fields on all possible
command option choices; additionally, these fields are all the same, and
the only difference is the type of the value. Instead of wrapping all of
the fields in an enum, we can make `CommandOptionChoice` a struct with
the `name` and `name_localizations` fields and an enum for *only* the
value. This looks like:

```rust
pub struct CommandOptionChoice {
    pub name: String,
    pub name_localizations: Option<HashMap<String, String>>,
    pub value: CommandOptionChoiceValue,
}

pub enum CommandOptionChoiceValue {
    String(String),
    Integer(i64),
    Number(f64),
}
```

This allows the fields equivalent across all value types to be directly
accessed.