docs(book): 0.15 release notes (#2074)
Add the release notes for 0.15 to the book. The release notes detail the
improvements made to the gateway, why they were made, and how to use
them. Also mentioned are notable quality of life changes and breakages,
such as the removal of `Member::guild_id`, support for retrieving guild
widget settings, the minor change to allowed mentions, and security
improvements to prevent debug formatting of tokens.