refactor: change deny lints to warn (#2144)
Development of Twilight can be somewhat tedious due to warning lints
failing compilations while working on something in progress. We should
instead only fail builds for lints once they're submitted as a pull
request and run through automated testing.

This removes the deny by default lints and warns on the remaining lints.
Additionally, an unnecessary `unused_underscore_binding` lint, which was
useful back when `tracing` was an optional feature.

This change was suggested in discussion #1644.