docs(model): clarify reaction type unicode values
Clarify that the name of a unicode reaction type
(`ReactionType::Unicode`) is the *value* of the unicode emoji and not
the Discord display name for the emoji. For example, instead of using a
value of "arrow_right", a name of "➡️" must be specified.

Documentation added after a support thread was received in our Discord
server:
<https://discord.com/channels/745809834183753828/1075631047871578153>