fix(gateway): ignore unknown event errors (#2154)
Ignore errors caused by attempting to parse events that are unknown to
Twilight. Because these are unknown events, they couldn't possibly ever
produce an event, and it's therefore expected that they fail to
deserialize. Because this isn't an *error* and considered a normal
operation in the event loop, we should skip over these errors from
unknown events.