feat(http)!: return errors at finalization
Refactor HTTP request builders to return errors when builders are
finalized and constructed into a `ResponseFuture` via `IntoFuture`.
Errors encountered during building are stored internally and will cause
`ResponseFuture`s to resolve to an `Error` of a new error type,
`Validation`.

This allows users to construct a request and not have to handle errors
at possibly several points during construction.

Closes issue #1187.