refactor(twilight-model)!: Un-`Box` `SelectMenuData` variants
This commit unboxes `SelectMenuData`'s variants. They were boxed in a
previous patch to prevent future fields from unnecessarily increasing
the enum's memory footprint, but considering that, with the current
structures, new fields are a breaking change already, it's OK to leave
the variants unboxed for now and re-box as necessary.

For the current code, this saves a heap allocation while only marginally
increasing the enum's memory footprint.

This change is only breaking in the context of PR #2219. When added to
the commits in this PR, it doesn't add any breaking changes to the PR's
list of breaking changes.