feat(model): remove `tracing` (#2238)
The `tracing` ecosystem of crates allow for recording events which can
then be viewed when debugging. Most `twilight` crates therefore record
useful data with it, but its unnecessary for `twilight-model` to do so.
This is because deserializers most necessary data on deserialization
error for troubleshooting. Including further details is left to
dependants. `twilight-gateway`, for example, includes the full gateway
event in `ReceiveMessageError`.