refactor(gateway-queue)!: rewrite crate (#2228)
Superceds and is a full rewrite of the proposed rewrite in #2172.
Primarily motivated by removing the use of a background task `Shard`
when waiting to identify. By returning a concrete `Future` type
(implementing `Debug`) it becomes much easier to store it inside of
`Shard`. Of course, I did not stop there, and you will find the
gateway-queue's documentation much improved in addition to code
reduction and an (initial) integration test suit. The provided
`InMemoryQueue` should be more flexible for advanced use cases than
`LargeBotQueue` whilst offering more features without added complexity
than `LocalQueue`.

Refer to #2172 for motivation on changing the identify interval from 6
seconds to 5 seconds.