feat(gateway)!: turn `Queue` into a generic type (#2249)
By making `Config`, and its dependants, generic over the queue we avoid
an unnecessary pointer indirection through `Arc<dyn Queue>` which has
been a pet peeve of mine for some time. This furthers allows dropping
the `Debug` requirement from `Queue`, another of my pet peeves.

As everyone not running multi-process bots should be using the
`InMemoryQueue`, `Config`, `ConfigBuilder`, and `Shard` default to it.
I.e. `Shard` is equivalent to `Shard<InMemoryQueue>`. This might,
however, not be ideal if further queues are added. For example, a
`SelfRefillableQueue` implementation that calls the "Get Gateway Bot"
endpoint in the background at some interval could prove very popular in
which case defaulting to any type might not be that useful.