feat(cache_in_memory)!: In-memory cache with generic types (#2179)
Adds support for using custom structs for cached representation of
Discord API data via generics on `InMemoryCache` and a set of traits for
cached data that is required for caching functionality.

The `DefaultInMemoryCache` type alias is added for an implementation
with the default types provided by the cache crate. Migration is very
straightforward, replacing `InMemoryCache` with `DefaultInMemoryCache`
in code is sufficient. 

This could potentially reduce memory usage drastically when structs
are optimized properly by implementors.

Signed-off-by: Jens Reidel <adrian@travitia.xyz>
Co-authored-by: Valdemar Erk <valdemar@erk.dev>