refactor(gateway): remove instrument from shard::poll_next (#2401)
Shard instrumentation is necessary for following specific shards, but by
adding it ourselves users are not able to hook into it, for example on
errors. Some work around this by instrumenting their shard running
function, but this may lead to duplication if they do not filter their
subscriber. This was considered worthwhile, as it cleaned up user's logs
at a time when many experienced issues with the gateway and
instrumenting a stream combinator (like the old `ShardEventStream`) is
difficult.

The `gateway-reshard` example needed to be migrated from using a stream
combinator to spawning tasks (which also improves its performance).