feat(model, validate, util)!: New modal components (#2461)
Label components [0] are the new recommended way to add 
components to modals.
Additionally, select menus and text displays are now allowed in
modals [1].
This PR adds label components as well as capabilities to use the new
modal components.

This also includes a change to `TextInput`: since labels should now be
added using the label component, the `label` field on text input
components are deprecated by Discord and now optional in the `TextInput`
type.

[0]: https://discord.com/developers/docs/components/reference#label
[1]: https://discord.com/developers/docs/change-log#adding-more-modal-components