feat(model, cache): add missing fields to Member (#2476)
Added fields to the [`Member`] struct which was missing in model.

https://discord.com/developers/docs/resources/guild#guild-member-object

permissions (ignored), banner and avatar_decoration_data was missing,
and in this PR it is added.

(permissions was ignored because we already have it present in
PartialMember with interactions, this is the only case where permissions
would deserialize to Some)