feat(lavalink)!: Lavalink refactor to v4 API (#2322)
Here are the changes required to work against the new lavalink API:
https://lavalink.dev/api/index.html.

I didn't fully implement all the API since some of it is new in `v4`. I
only maintained the compatibility of the current system that was built
on `v3`. I put the structs / enums under deserialization and
serialization tests with json i received when testing the bot and with
some in the documentation. I also tested against my bot and seems to
work just as well as before.

Major changes were:
* REST API now for all outgoing events. I used a hyper http client for
this.
* Websocket is only used for connection and incoming
* Structs and enums changed names and updated members
* New api now versions the api in the uri `/v4/...`
* V3 lavalink servers are not compatible now
* Refactored all of the incoming and outgoing into their own files
within model folder

There are a few other things i want to change but they aren't critical.
I want to clean up the http since it has some model structs that could
make more sense in model. The http also just generates the body and
parts for the http requests for some endpoints. I think the player needs
to hold the client and make these calls for the user. There is too much
detail exposed in my opinion. That will be a different day and effort
but wanting to jot that down now.

Please let me know what I need to change and can do those and test with
the basic bot. I updated the documentation with what i thought was
relevant.

This covers the issue: #2192